A Gatsby starter for Shards UI & Web Components
I've been using Shards UI for 3 of my projects now, and I was getting really tired of repeating myself every project. To avoid this redundancy I created a special template that uses all of the standards to start your own Gatsby Project. Those are typescript, styled-components, and a UI framework.